refactor(optimiser): move type definitions to a separate file for better organization and update imports across components
